fix: made the name field optional on Class type
The created Typescript classes has no name property, but the compiled Javascript object has.  This change allows to keep the structure of the classes and there is no need to define a name property.

Added a check in the buildModel and buildInput functions to ensure safety, however they should never be triggered, as long as a proper class is passed to the builders.